% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ci.R
\name{calc_ci}
\alias{calc_ci}
\title{Calculate confidence intervals for lm, glm, and svyglm}
\usage{
calc_ci(model, conf_level = 0.95, method = c("wald", "profile"), df = Inf)
}
\arguments{
\item{model}{The model/data object}

\item{conf_level}{The confidence interval, default is 0.95}

\item{method}{The method for calculating CIs}
}
\description{
This function calculates confidence intervals for \code{lm}, \code{glm},
\code{svyglm}, and \code{svystat} objects. The \code{svyglm}and \code{svystat} methods
are based on code from the\code{survey} package. In addition, there are
two methods: wald and profile likelihood.
}
\keyword{internal}
