% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Clean and augment model output}
\usage{
clean(
  model,
  conf_level = 0.95,
  include_reference = FALSE,
  conf_method = c("wald", "profile")
)
}
\arguments{
\item{model}{A fitted model object (e.g., \code{lm}, \code{glm}, or \code{svyglm}).}

\item{conf_level}{Confidence level for intervals (default = 0.95).}

\item{include_reference}{Logical; whether to include reference levels for
categorical predictors.}

\item{conf_method}{Determines whether the confidence intervals are calculated
using the profile likelihood or the Wald method. Obviously has two options,
"profile" and "wald". Wald is between 3 to 25 times as fast but not as
reliable for small sample sizes (n < 50). For larger sample sizes,
(n > 100), they will be very similar. \strong{The default is Wald.}}
}
\value{
A tibble of coefficients with confidence intervals and counts.
}
\description{
A simplified alternative to \code{broom::tidy()} that adds confidence intervals,
optional reference levels, and sample sizes for each term.
}
\keyword{internal}
