% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_svyglm.R
\name{fit.svyglm}
\alias{fit.svyglm}
\title{Calculate R-Squared and Adjusted R-Squared}
\usage{
fit.svyglm(svyglm, decimals = 3)
}
\description{
Based on McFaddden's pseudo-R-squared. This shows how good of a
fit the model is. \strong{NOTE: It does not show the percentage of
variability explained by the model}
}
