% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_col_names.R
\name{get_col_names}
\alias{get_col_names}
\title{Get the column names}
\usage{
get_col_names(data, cols)
}
\arguments{
\item{data}{a \code{data.frame} or \code{tibble}}

\item{cols}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> One or more unquoted
expressions separated by commas. Variable names can be used as if they
were positions in the data frame, so expressions like x:y can be used
to select a range of variables.}
}
\value{
a vector of columns in data
}
\description{
This function is essentially a wrapper around \code{dplyr::select()} and \code{colnames()}.
\code{get_col_names(data, cols)} is the the same thing as
\code{colnames(dplyr::select(data, cols))} just with slightly less typing.
}
\examples{
library(tidysurvey)
data <- make_basic_df()
# let's get all of the column names between "grp" and "satisfaction_service"
get_col_names(test_data, c(grp:satisfaction_service))
# can also use tidyselect syntax
get_col_names(test_data, tidyselect::starts_with("satis"))

}
