% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_longer_values.R
\name{pivot_longer_values}
\alias{pivot_longer_values}
\title{Pivot data from wide to long with value labels}
\usage{
pivot_longer_values(
  data,
  cols,
  names_to = "names",
  values_to = "values",
  name_label,
  ...
)
}
\arguments{
\item{data}{A data frame to pivot.}

\item{cols}{<\code{\link[tidyr:tidyr_tidy_select]{tidy-select}}> Columns to pivot into
longer format.}

\item{names_to}{A character vector specifying the new column or columns to
create from the information stored in the column names of \code{data} specified
by \code{cols}.
\itemize{
\item If length 0, or if \code{NULL} is supplied, no columns will be created.
\item If length 1, a single column will be created which will contain the
column names specified by \code{cols}.
\item If length >1, multiple columns will be created. In this case, one of
\code{names_sep} or \code{names_pattern} must be supplied to specify how the
column names should be split. There are also two additional character
values you can take advantage of:
\itemize{
\item \code{NA} will discard the corresponding component of the column name.
\item \code{".value"} indicates that the corresponding component of the column
name defines the name of the output column containing the cell values,
overriding \code{values_to} entirely.
}
}}

\item{values_to}{A string specifying the name of the column to create
from the data stored in cell values. If \code{names_to} is a character
containing the special \code{.value} sentinel, this value will be ignored,
and the name of the value column will be derived from part of the
existing column names.}

\item{name_label}{Add a variable label to the new column with the names of
the columns}

\item{...}{Additional arguments passed to \code{\link[tidyr]{pivot_longer}}.}
}
\value{
A "long" data.frame.
}
\description{
This function is a wrapper around \code{\link[tidyr]{pivot_longer}}. This
function operates in pretty much the exact same way but uses the variable
labels from the variables specified in \code{cols} to make new value labels in the
new variable created in the \code{names_to} variable.
}
\details{
An additional note is that this function also works with survey objects
created with either \code{survey::svydesign()} or \code{sryvr::as_survey_design()}.
The function first pivots the data, then re-creates the survey object
using the same variables used
}
