% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_longer_values.R
\name{pivot_longer_values_core}
\alias{pivot_longer_values_core}
\title{pivot_longer_values main interal helper function}
\usage{
pivot_longer_values_core(data, cols, names_to, values_to, name_label, ...)
}
\description{
Takes a data frame, and pivots it long-ways while preserving
label attributes.
}
\keyword{internal}
