% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_groups}
\alias{select_groups}
\title{Select variables from the group variable}
\usage{
select_groups(group, data)
}
\arguments{
\item{group}{\code{tidy_select} columns to group the data by}

\item{data}{The data set that the columns are from (this to make sure they exist)}
}
\description{
Select variables from the group variable
}
\keyword{internal}
