% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars_pval.R
\name{stars_pval}
\alias{stars_pval}
\title{Add stars based on the p-value}
\usage{
stars_pval(p.value)
}
\arguments{
\item{p.value}{A vector of p-values to determine what stars to give}
}
\description{
Add stars based on the p-value
}
