% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidysurvey-package.R
\docType{package}
\name{tidysurvey-package}
\alias{tidysurvey}
\alias{tidysurvey-package}
\title{tidysurvey: Conduct survey analysis using tidysyntax}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Jacob Dennen \email{jdenn0514@gmail.com}

}
\keyword{internal}
